application/src/main/kotlin/com/stringconcat/dev/course/app/event/EventPublisherImpl.kt
package com.stringconcat.dev.course.app.event

import com.stringconcat.ddd.common.types.base.DomainEvent
import com.stringconcat.ddd.common.types.base.EventPublisher

class EventPublisherImpl(listeners: List<DomainEventListener<out DomainEvent>>) : EventPublisher {

    private val listenerMap = listeners.groupBy { it.eventType() }

    @Suppress("UNCHECKED_CAST")
    override fun publish(events: Collection<DomainEvent>) {
        events.forEach { e ->
            listenerMap[e::class]?.let { listeners ->
                sendEvents(listeners as List<DomainEventListener<in DomainEvent>>, e)
            }
        }
    }

    private fun sendEvents(listeners: List<DomainEventListener<in DomainEvent>>, event: DomainEvent) {
        listeners.forEach { l -> l.handle(event) }
    }
}